Support multiple queries with JSONP.
Since there are complexities with representing an array of queries (objects)
in a querystring, an additional API entrypoint has been added.

The new entry point is the q= parameter in the querystring.

q expects a url encoded json blob of one or more queries as
documented in the README.

This is the recommended approach for people wanting to make multiple
queries in one request via JSONP.

A querystring representation of a single querystring is still valid.
People should use this if they are on a client which cannot
stringify JSON. However they should note that they can only make one
query per request.