Fix inconsistency with error reporting.
The problem with the error reporting being inconsistent, ie
only one result object showing despite 2 or more queries being
sent is a misunderstanding with the caolan's async#parallel
function.

Mistakenly I was passing the error response to the callback for
the asynchronous tasks. This was a mistake as one error cancels
the rest of the parallel tasks. Meaning the final callback comes
early.

To fix I removed the error being passed into the callback for each
asynchronous task. Despite errors, each task gets a chance to be
completed before the final callback is called.